mnormt-package {mnormt}	R Documentation

The 'mnormt' package: summary information

Description

This package provides functions for computing the density and the distribution function of multivariate normal and multivariate Student's t variates and for generating random vectors sampled from these distributions.

Details

Probabilities are computed via a non-Monte Carlo method. Different routines are used in the three cases d=1, d=2, d>2, if d denotes the number of dimensions.

Licence

This package and its documentation are usable under the terms of the “GNU General Public License” version 3 or version 2, as you prefer; a copy of them is available from http://www.R-project.org/Licenses/.

Author(s)

Adelchi Azzalini (R code and package creation) and Alan Genz (Fortran code, see references below; this includes routines of other authors)

References

Genz, A. (1992). Numerical Computation of Multivariate Normal Probabilities. J. Computational and Graphical Statist., 1, 141-149.

Genz, A. (1993). Comparison of methods for the computation of multivariate normal probabilities. Computing Science and Statistics, 25, 400-405.

Genz, A.: Fortran code available at http://www.math.wsu.edu/math/faculty/genz/software/fort77/mvn.f

[Package mnormt version 1.5-5 Index]